+++ src/test/java/io/github/astrapi69/io/StreamExtensionsTest.java
@@ -46,7 +46,6 @@ import org.junit.jupiter.api.Test;
 import org.meanbean.test.BeanTester;
 
 import io.github.astrapi69.date.CreateDateExtensions;
-import io.github.astrapi69.io.shell.LinuxShellExecutor;
 import io.github.astrapi69.lang.ClassExtensions;
 import io.github.astrapi69.test.base.BaseTestCase;
 import io.github.astrapi69.test.object.Person;