+++ src/test/java/io/github/astrapi69/io/file/FileExtensionTest.java
@@ -46,8 +46,25 @@ class FileExtensionTest
 	{
 		String actual;
 		String expected;
+
 		actual = FileExtension.PROPERTIES.getExtensionOnly();
 		expected = "properties";
+		assertEquals(expected, actual);
+
+		actual = FileExtension.JSON.getExtensionOnly();
+		expected = "json";
+		assertEquals(expected, actual);
+
+		actual = FileExtension.XML.getExtensionOnly();
+		expected = "xml";
+		assertEquals(expected, actual);
+
+		actual = FileExtension.YML.getExtensionOnly();
+		expected = "yml";
+		assertEquals(expected, actual);
+
+		actual = FileExtension.YAML.getExtensionOnly();
+		expected = "yaml";
 		assertEquals(expected, actual);
 	}
 
@@ -93,7 +110,7 @@ class FileExtensionTest
 		Set<String> actual;
 		Set<String> expected;
 		actual = FileExtension.getZipExtensions();
-		expected = new HashSet<>(Arrays.asList(FileConstants.ZIP_EXTENSIONS));
+		expected = new HashSet<>(Arrays.asList(FileExtension.ZIP_EXTENSIONS));
 		assertEquals(expected, actual);
 	}
 }