+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 ## Change log
 ----------------------
 
+Version 3.3
+-------------
+
+ADDED:
+
+- new class CommandLineExecutor for executing shell commands across different operating systems
+- introduce OS enum for OS detection and integrate it with CommandLineExecutor to set appropriate shell paths and flags
+- new Implement factory method for creating ProcessBuilder instances to simplify process configuration
+
+CHANGED:
+
+- remove of deprecated class FileConstants
+
 Version 3.2
 -------------